refactor(update-server): Split `name_management` into submodules (#10523)
