+++ update-server/tests/common/name_management/test_pretty_hostname.py
@@ -1,6 +1,8 @@
-from otupdate.buildroot import name_management
-import pytest
 from collections import Counter
+
+import pytest
+
+from otupdate.common.name_management import pretty_hostname
 
 
 machine_info_examples = [
@@ -14,7 +16,7 @@ machine_info_examples = [
 @pytest.mark.parametrize("initial_contents", machine_info_examples)
 def test_rewrite_machine_info_updates_pretty_hostname(initial_contents) -> None:
     # TODO(mm, 2022-04-27): Rework so we don't have to test a private function.
-    rewrite = name_management._rewrite_machine_info_str(
+    rewrite = pretty_hostname._rewrite_machine_info_str(
         initial_contents, "new_pretty_hostname"
     )
     assert (
@@ -29,7 +31,7 @@ def test_rewrite_machine_info_updates_pretty_hostname(initial_contents) -> None:
 def test_rewrite_machine_info_preserves_other_lines(initial_contents) -> None:
     # TODO(mm, 2022-04-27): Rework so we don't have to test a private function.
     initial_lines = Counter(initial_contents.splitlines())
-    rewrite_string = name_management._rewrite_machine_info_str(
+    rewrite_string = pretty_hostname._rewrite_machine_info_str(
         initial_contents, "new_pretty_hostname"
     )
     rewrite_lines = Counter(rewrite_string.splitlines())
@@ -44,10 +46,10 @@ def test_rewrite_machine_info_preserves_other_lines(initial_contents) -> None:
 @pytest.mark.parametrize("initial_contents", machine_info_examples)
 def test_rewrite_machine_info_is_idempotent(initial_contents) -> None:
     # TODO(mm, 2022-04-27): Rework so we don't have to test a private function.
-    first_rewrite = name_management._rewrite_machine_info_str(
+    first_rewrite = pretty_hostname._rewrite_machine_info_str(
         initial_contents, "new_pretty_hostname"
     )
-    second_rewrite = name_management._rewrite_machine_info_str(
+    second_rewrite = pretty_hostname._rewrite_machine_info_str(
         first_rewrite, "new_pretty_hostname"
     )
     assert second_rewrite == first_rewrite
